feat(frontend): Add kubernetes setup
I have re-worked an example jsonnet data template to work with orbis.
We need to generate the kubernetes config, from this jsonnet template,
to create:

* deployment.json
* service.json

at the minimum (although I also have a `privatepods.json` to manage
access to our private AWS ECR registry. There is also a new GitHub
action to do all this in an automated manner.