fix(frontend): Fix creation and deployment of production artifact
I have now fixed the actions involved in building, testing and more
importantly, the deploying of the app to a kubernetes cluster. This
fix included updating the jsonnet config to swap branch name to tag
GIT HASH (first 7 chars) as well as using the correct `Dockerfile.artifact`
rather that the development one.