fix(project): removed references to GIT_REF
Removed references to $GIT_REF in the build action.  This had been
referencing "master", but the cached version was always being pulled b/c
it was still cached from previous builds.  Instead we only ever tag an
image w/ the $GIT_SHA (which changes on each push and will therefore
never be cached).