fix(frontend): Fix map not displaying
I made changes to remove a hooks warning, not realizing that I broke
the map. The effect was only being run the first time it was called,
not when parameters changed, or at least the right ones. I've now
fixed it so that when the mapbox token is retreived from the server,
this effect is re-run.