fix(project): Re-factor docker-compose
It has been decided for development setup, to separate the client/server
into 2 containers, this makes it quicker for the backend developer(s).
This meant removing references to the client in the `Dockerfile` and
`run-server.sh` script. I also had to update the `proxy` line in the
client's `package.json` file, so API requests proxy through to the
server correctly