fix(frontend): Set background image separately
I initially set the background to an image for all forms, at the time
this was correct as only the `user account` forms existed. Now there
are other forms and I want them to also re-use as much styling as
possible, so refactored out the setting of the background to another
CSS class and added that to each relevant form.