fix(frontend): Remove action parameter for user errors
I was dispatching the actual error when calls to get the current
user failed. I expect this call to fail, so no point raising the
error (I may decide differently at a later date though), especially
as it is displayed on the login/register/forgotton password pages,
which we don't want. Removing the action was just the simplest solution.

I now no longer pass the actual username/email in the `fetchUser` call
as `current` is good enough. I may revisit this when we re-add the
`Admin Panel` though.

I fixed the form element name in the `Password Reset Confirm` form.

I removed obsolete `history.push('/logout')` as the route no longer
exists