feat(frontend): Make mini-map more zoomed out
The mini-map is to give an overview of where you are, so it needs
to be more zoomed out than the main map. The problem was the size
of the map's canvas was bigger than the visible area, so it was
centered correctly. I've added some code to re-size the map once
loaded, which makes the map only take up the space wanted.