feat(frontend): Make Change Password form part of sidebar
Added ability for the user to change their password. This is only the
beginning, it is not yet finished. The functionality of this is that
the change only becomes `live`, once the user clicks a verification
link in the email sent. We still need to handl this on the front-end
and display a success/error message.