fix(frontend): Fix toolbar open/close functionality
Passed toolbar item label into toggleMenu function and corresponding action (matching toggleMenuItem), so that reducer has access to it.

Then added a conditional to check if, once open, the newly clicked item is the same as the currently displaying item. If so, the sidebar simply closes. If not, the sidebar remains open and updates the information displayed to match the clicked item.

IssueID #