fix(frontend): Allow user to select area of interest
In the `Landing Page`, the button opens a dialog to allow the user
to select domain and region to configure the map. I have now made
this code work so you can get through to the map easily and it is
properly configured.

I feel how we make the map appear in a certain region is going to
have to change. I expect us to what to do `intersections` e.g. This
region polygon intersects with these domain polygons, so make those
domain polygons available to users. This cannot be done with this
more simplistic solution.