fix(frontend): Re-style annotations again
This one is a pain and couldn't be foreseen. The existing styling
worked fine in the development environment, as all CSS is loaded in
`style` blocks. Whereas when run in production, the generated build
creates application code as links to css and inlines vender (including
our UI Library) after, meaning the vender styles load after our customisations.
We need to force the issue, but can only prove it by viewing the
production build, hence the annoyance.