feat(frontend): Integrate Data Server
We now need to start getting real data sources, these come from
the API. I periodically query the API to get a new token (it expires
after every one hour by default). We actually replace the token
every have that value, so if it is one hour, it is actually refreshed
every 30 mins).