fix(frontend): Fix deletion of bookmarks
The deletion of bookmarks was broken as we didn't pass the user's
unique key in the request, having now done that, we can delete again.

I commented out the request to save bookmarks as the new method has a
dependency on code being written on the backend, not sure when that
will get done, so commented out the code for that right now and reverted
back to submitting a JSON object rather than a FormData one.

IssueID #57