fix(frontend): Check if requestIdleCallback exists
Really not sure what is going on here, we use `window.requestIdleCallback`
function, in this and other apps, but for some reason in this app, Safari
claims the function doesn't exist. The function removes the map, which doesn't
happen a lot I think, but not exactly sure either. Everything works as it
did before for all browsers except Safari, we will need to keep an eye on
it.