fix(frontend): Disable notifications prior to login
When a user visited the app, prior to login, they were seeing an
error notification informing them there was an error retrieving
data sources. This doesn't make sense before the user has even
logged in. I have now added code so the request is only made if
the user exists. I've commented out a few other instances that
could cause issues during the demo.

We do need a mechanism to inform the user when there are problems,
what it is and how to recover, if possible. Notifications were a
simple mechanism for this.

IssueID #81