fix(frontend): Fix sync'ing of map location in url
A bad situation happens when first going to `/map`, there is no viewport
object in the URL, so the default map location is removed altogether. I've
now made it that state is only updated if a viewport object exists. There
are a few ways we can do this, this works, but it could easily need to change
in future again.

IssueID #82