fix(frontend): Fix LoadMask errors
It turns out the loadmask issue was more complicated than initially thought.

I thought since it was a simple true/false toggle, that local state inside the map component could be used.
However, the fact that the bookmark selected by the user is stored in global state caused an error when the same bookmark is clicked twice.

The only recourse was then to do the same as before, but using global state, and short-circuiting the action dispatch so that it is not run in the event that the clicked bookmark is the same as the one currently held in global state.