fix(frontend): Fix landing items spacing
I was never happy with the layout of the landing page, so took this opportunity to remedy it.
I have removed the padding on the sides of the landing page content container, which was forcing the width, and just aligned the entire content container to the center using flex. This allows the page to reduce until there is no more whitespace before adding the scrollbars. I think this is an appropriate solution for the short-term.

I also added a right-margin to the items, so that they can never touch, even if space runs out between them.