fix(frontend): Continue fixing styling and cleaning up code
I was never happy with the landing page 'View All' layout, so I took this opportunity to refactor the component and fix some of the styling. Of course, not all of this can be dealt with now, while the design team are busy, but I have removed the passing of the classnames as props as I felt this was bad practice and generally ugly. I have switched the 'View all' to flexbox from grid, and removed the title/setTitle state, as well as all it's references. It wasn't funtioning anyway, so I have just hardcoded a title until I hear from the design team what is to be displayed there.