fix(frontend): Import new icons, adjust styling/positioning
I imported the new icons from Zeplin, ran them through svg.io and styled the so they are the correct proportions. The 'T' icon for the 'text' button (annotations panel) id apparently supposed to be a literal 'T' character, so the icon was ditched in favour of this.

I also commented out the 2 icons and their functionality (rotate and upload image), but did not delete them. This was by instruction, as they do not feature in the designs.

Lastly, I replaced the 'Add Category' icon with the proper icon.