fix(frontend): Adjust minor styling
I adjusted some of the styling on the side-bar and spaceman landing page. The landing page was just to add a bit of space between the buttons (to stop them touching), and the side-bar was to fix some small bugs brought in by library changes and also to match deisgns a bit better.