fix(frontend): Refactor to fix tests
Before the code worked but the tests for the MapStyleSwitcher kept
failing due to redux code. To fix this, instead of passing the
actionCreator function selectMapStyle to the switcher component and
then dispatching the function, I now wrap the dispatching of the
selectMapStyle function inside another function and pass it to the
MapStyleSwitcher. This allows jest mock function to work again as it
had done before.

I moved the MAP STYLES to the `.env` file instead of `.env.local` as
they are neither secret or developer dependent.

I set the default style to be `satellite` as we have used previously.

I also simplified the MapStyleSwitcher, remove unnecessary return statements.

IssueID #157