fix(frontend): Downgrade versions of semantic release dependencies
Not sure why this is happening, I had upgraded these and re-installed
having first deleted my node_modules and it worked, but no it isn't.
No matter, I have downgraded the versions to the latest before they
need NodeJS 10.18.1 or above (we are using 10.16.1), likely due to
some other dependency as we specify the Node version to be 12.13.1.