fix(frontend): Integrate frontend/backend satellite search
Until now we have worked on the UI and API in parallel. The UI had
a development API to indicate the sort of response from the API.
Now the API had had more development it was time to integrate. This
task is not yet complete, we are getting errors from the Orbis API.

We have experienced a problem in that we define a `Polygon` for the
AOI, whereas we think the Copernicus API uses `MultiPolygon`.