fix(frontend): Update info icon and time format
(I saved too early on the previous commit, this one explains fully).

I imported the info icon from the ui-library instead of from directly within the app, and deleted the SVG of the previous icon.
I also restructured the time format on the 'scenes' metadata to use the satellites constants to match the ComparePins component.