fix(frontend): Re-add info SVG
In the previous commit, I pulled in the InfoIcon from the ui-library, and since it was un-needed, deleted the SVG file from the app.
As this was late in the day, I overlooked the fact that it was being used in other places too (Search/ComparePins).

This left 2 choices: 1) Update the icon in the other panels to match this one, or 2) replace the SVG and remove it as a separate issue.

I chose to replace it for now, as this branch is only concerned with results and should not be venturing into other panels.