fix(frontend): Change default mapstyle
We used thermcert's custom mapstyle but it was overwritten almost
immediately. I am changing it as the thermcert one no longer exist,
since we change it immediately, we never noticed.

At some point we will need to use a specific one for orbis as we will
want some custom changes to the mapstyles.