fix(frontend): Refactor rendering of icons in scene items
I updated the scene list item component to render the pin/delete icons in a more efficient way.

The scene list items are virtually identical on the results and compare tabs, with the exception of the icon.
Previously, there was a string prop and if/else to determine which icon to use, but that involved duplicate code.
Now, each tab (results/compare) simply has a renderProp function that passes in its icon and inside the component, that function is simply called.