fix(frontend): Refactor multiples modals/dialogs into single instance
There was a total of 4 separate modals and dialogs across the search/results and compare panels, each using local state, the ui-library's useModal hook and code for the modal/dialog itself.

I've gutted out all of that and replaced it with a single modal/dialog in the parent component (satellites-panel), and renamed the variables/functions with generic names, so that they are re-used across all instances.