feat(backend): added maximum AOI area setting
Added maximum AOI area setting as a `DynamicSetting`.  This value can be
changed w/out a code change.  And it gets passed to the frontend as
part of the `/config/` endpoint.

There will still be work to do to actually prevent a query from being
run if the AOI is greater than this value.

IssueID #193