fix(frontend): Fix scrollbar issue
A while back, before the Malaysia demo, there was an issue with scrollbars appearing when moving the map. As a quick fix for this, I hid the overflow in the body of the app, which worked, but also removed scrollbars everywhere.

I have remove that, and the scrollbars are back as desired, but the undesired ones on the map have not also returned.
My guess is that this is due to the MapBpx styles issue being resolved, and may not be a problem needing fixing.