fix(backend): fixed error saving result
Saving a result (pinning a scene) was failing for two reasons:

* the scene_id was sometimes too long
* the footprint couldn't be serialized

Fixing the scene_id issue was easy; I just added a max_length kwarg to
the field (and updated the db schema in a migration)

Fixing the footprint was harder.  Because the satellite adapter could
return either `Polygons` or `MultiPolygons` as the footprint, I used the
generic `GEOSGeometry` field on the model.  The adapter works w/ the WKT
format of a geometry and saves it directly to the model.  But the
serializer (used when saving a result) was using a custom
SimplifiedGeometryField which just returned an array.  Trying to
deserialize that array back into a `GEOSGeometry` object was failing
(b/c it expected the WKT).  To get around this, I used the built-in
DRF-GIS serializer which works with GeoJSON.

In the future, I may revisit this and write some code to go from an
(arbitrarily-nested) array to WKT.  I suspect that working with arrayns
in the frontend is simpler than working with GeoJSON once it has to
start doing things like filtering by bounding box.  In the meantime,
though, this is good enough.

IssueID #215