fix(frontend): Build 'more info' modal data table structure
The modal was getting very messy when attempting to make it generic for any information to be passed in.

The reason being that some of the information would just be simple key/value pairs, but for some, the values would be objects or arrays, which themselves may contain objects or arrays. Attempting to grab the keys, then use them to map the various fields, then have conditionals in place to catch the objects/arrays, and map them out again, and so on, the modal was becoming spaghetti code of maps and loops.

In order to get around this, I built 3 separate table components, one for each use case, in a separate file and imported them to the panels where they are required. I then pass the data (scenes/satellites/tiers) as a prop, and pass the entire component as the infoContent.

On the modal, if there is content, the table component is rendered. If not, a paragraph of the text 'No information currently available' is displayed.

To me this appears to be the cleanest way and mort non-repetitive way to do this. And the added bonus is that the info dialog can now be re-used anywhere else in the application, which it may need to be somewhere down the line.