fix(frontend): Added split map scene display functionality
I added functionality to display 2 scenes on the 2 different maps during compare mode.
When the 2 maps are rendered, the 2 scenes for comparison are also passed in.

The checkboxes are now disabled during compareMode, although I'm not sure if this is the intended solution. I did this because de-selecting the scenes while still in compareMode is problematic.

Toggling off compareMode removed the sources and layers, in the same manner as the visualisations panel.

Also removed some console.logs.