feat(frontend): Authorize access to features
We want the richness to be able to authorize users access to feature
components by `roles`. The backend already supplies a list of roles,
along with the user. So, we can use the `useAuthorization` hook to
check the user is authorized to use these features.

The one thing of note is that the client will need to know exactly
what `roles` are permitted to use it, so we can't turn that on/off
without code changes. We will have to see how maintainable this
solution is.

I've added an example of authorization to the toolbar as an example.
This should work just fine with the express dev server and the django
one.