WIP feat(frontend): Add popup when Hospital map feature clicked
This code extends what was already there, to display a popup when
the Hospital feature is clicked. The data displayed relates to that
hospital. If the feature is a cluster of hospitals, then the map
zooms in and re-centers, re-displaying features.

Features are either `red` for cluster, or `green` for individual, that
when clicked, displays the popup.

This was taken from thermcert, so the detail is very specific to the
Hospitals. This is going to be a common theme, we want different metadata
displayed in the same Dialog. We need to make the component displayed
inside the popup, be more adaptable to what it is to be displayed,
no data will have the exact same metadata.