fix(frontend): Added cleanup to useEffect, returned compare image back to original in proxy
I returned a clean-up function from the useEffect of the comPare Pins panel.
The reason for this is that the clean-ups in the useMap for the comparing of pins removes the layers and sources, but only when compareMode is exited, not whent the compare component unmounts. The clean-up in the compare component simply toggles compareMode off, which removes the cources/layers.