fix(frontend): Fix covid modals
I now realise that the modal for multiple users is just the one for a single user with many being iterated over.
So to fix this, I had to map lal of them into the function that then maps the list elements.

I also took the scrollbar from the Satellites panel and re-used it here.

I had to set a min-width to the modals sop they are not cramped and match the designs.