feat(frontend): Switch redux store to use @reduxjs/toolkit
We have been wanting to switch to @reduxjs/toolkit for some time.
This is the way the creators of redux always intended developers to
use redux as it reduces boiler-plate code, which redux has lots of.

So far I've only updated the store to use the toolkit, more commits
will migrate existing `actions`, `reducers` etc to `slices`.