fix(frontend): Update how we use redux-persist
I'm not sure where I got the original code to integrate redux-persist,
but it is different from the website, likely just out of date. To
that end I've updated the code to match what is in their docs.

I'm doing this as I've been having problems integrating the new
reduxjs/toolkit slice, for the `accounts` actions/reducer.