fix(frontend): Update scene more info
The code to display the info was breaking the app as the data shape
was different from what was expected. I'm not sure which is more
correct, need to discuss, but I've made it work for right now. If
the old way is right, we can revert this commit and fix the API.