fix(frontend): Fix Map Style switcher
We switched from the mapstyles being hard-coded in the map redux state,
to getting the mapstyles as config from the server. We didn't update, or
lost the change to get the list of mapstyles from the right bit of state,
I've now updated this.

I'm not sure about this though, it does mean the state is part of the
app config, but the map has the actions and retains the current style.

IssueID #322