fix(frontend): Harmonize styling of popups
It is not going to be possible to guarantee a single Health Infrastructure
on one point, so we should be more flexible and treat it the same way
we do the Population Information, where we display the info for all features.
Since the styling should be the same, I factored out the selectors into a
shared module and import it into both components.

Before, we only passed the first feature, no matter, now we are passing
however many there are.

IssueID #327