fix(frontend): Fix checkboxes password reset/change form
The 2 forms for resetting/changing password had the same error as the Register form, in that a user can submit the form without checking the terms and conditions checkbox.

Adding a piece of local state to check if the box has been checked before enabling the button fixes this, the same fix used for the Register form.