fix(frontend): Update missed URL, remove brackets, fix import
I missed a URL, it is now added.
I also had to remove some brackets, as the constant was being read as an object.
Finally, I moved had to change the import in the PrivateRoute component to '.accounts/accounts.constants', as it is in a different directory from all of the others.