fix(frontend): Fix test after Updating dependency lock file
Having rebased the `master` branch, due to conflicts, I deleted and
re-created the `yarn.lock` file. This seems to have caused an knock-on
effect to the unit/integration tests, which I have now fixed. The
problem is, when using `getByLabelText`, on our `Switch component`,
we get both the input and button returned. We probably only want
to return one, but for now that isn't the case. If we consider this
a bug, we should fix it separately.