fix(frontend): Change how we get ENV variables
Initially I just added ENV variables to my `.zshrc` file, but that
isn't really the right thing to do. I have now changed the system
to use a `cypress.env.json` file. Each developer will have to create
their own as the file will not be committed to the repo.