fix(frontend): Make cypress work with absolute imports
While we have used the recommended solution from Create React App,
to enable absolute imports, the introduction of cypress testing has
thrown a bit of a spanner into the works.
I have now added an ENV variable to the script that runs the tests
to work-around this. This is not our preferred solution though, really
we would like to use a jsconfig.json instead, but having tried that,
it doesn't work, so this is an acceptable fallback I feel.