fix(frontend): Re-create yarn.lock file
I deleted and re-created the yarn.lock file in the hope it will fix
the build issue we are seeing in GitHub actions. Basically it fails
and I think it is down to there being multiple versions in the lock
file. Having re-generated, I don't see the same duplication, so here's
hoping.