fix(frontend): Fix warnings for satellite searches
I've committed this separately as it was more involved to fix and
needed a little more thought to ensure it didn't break anything.
Basically a couple of functions change on each render, so had to
memoize them by wrapping them in a `useCallback` hook. I've tested
the functionality afterwards and see nothing breaking, so think this
is fine.