feat(frontend): Update map to display satellite scenes
There was a few things wrong with the previous setup, for instance,
the individual scene to pass to each map was in fact changed to be
a boolean, not sure why, but that was wrong. Also, clicking compare
got into some infinite loop and ended up crashing the app, so I changed
the hook dependencies so it is only run once.

The map layout tset was also failing since I changed that component
so added the necessary state so the test passes again.