fix(backend): Update the URL for satellite scenes
I've updated the URLs used for satellite scene thumbnails and tiles.
I removed the HTTP prefix as this will differ depending on us running
the OLSP locally vs in `testing`, `staging` etc.

I uppercased the VISUALISATION_ID, for no other reason as it makes what
is to be replaced stand out a bit more.