fix(frontend): Add functionality for when you select a different visualisation
Although we only have the ability to use one visualisation right now
(True Color), I have added the ability to set the visualisation value.
Right now I don't know what the visualisation IDs will be, I only know
this one, so I've hard-coded it so that whatever one selected, it sets
it to `TCI`, as in True Color.