fix(frontend): Fetch pinned scenes when Satellite panel opened
I've moved the code to fetch already pinned scenes to the parent
component as more than one child requires this to have been done
before they can work. I previously had it in the ComparePins component,
but that caused an exception if you try to pin a file before having
gone to the ComparePins view. This is desired functionality, so we
need to know what scenes have already been pinned, so we can display
which have already been as well as actually pinning scenes.

IssueID #374