fix(frontend): Prevent deleting Pinned Scenes if selected
To avoid adding complicated code to manage various situations where
we could delete a pinned scene that is selected for viewing. I have
just disabled the `delete` button when the scene has been selected.
This is far simpler and avoids a number of potential bugs due to
complicated interactions.

IssueID #375