fix(frontend): Update landing page
There was actually more to this than I thought. First off, the modal and all of its props/functionality, and forwarding of ref was still in the Spaceman landing page, so that all had to be removed.

Secondly, because the previous image was an SVG inside the landing component, but the new image is a background for the entire component with the content on top of it, this caused a few issues having to separate out some styles that were shared bewtween the 2 landing pages.