fix(frontend): Fix Firefox error and toolbar width, remove props from landing page
The toolbar in Firefox was broken, adding a fallback argument for width fixes it.

Also, the "invalid token" error is still hitting in Firefox, so a check in the accounts.slice erorTransformer function blocks it.

And the unneeded props in the landing page (the ref was causing a warning) have been removed.