fix(frontend): Fix resendEmail error
Turns out it was a front-end issue. The function that formats the headers looked for a token that wasn't there. This wasn't a problem when registering, as there was a token - so this error went unnoticed.

When attempting pre-login, the token is not present, so the endpoint returns an error.

Changing this to the JSON_HEADERS constant we already have, it no longer looks for a token, and returns a 200 response.

(NOTE: the email had to be formatted from a string to an object in order to conform to the endpoint's spec).