fix(frontend): Fix buttons in T&C component
The lower opacity on the active button was making the buttons appear to inactive when they are in fact active.

This is being caused by the library, so this is the only way to fix it inside the app. I believe the buttons/links should all be the same app-wide for ease of use, so a decision will need to be made to determine if this is a change that is wanted for all buttons/links.