fix(frontend): Display correct legend for data
The code around selecting the legend is fragile as it relies on
hard-coding the name of the layer or part of it, in code. I have
stuck with this for now, but think we will need to have structure
to the naming of layers, gets really scary if they want 3rd party
data sources!!

Anyway, this just adds the missing new infrastructure we've added to
the list that was already there.

IssueID #416