fix(frontend): Refactor landing page
This wasn't really a pressing issue until recently, but the landing page's "View All" layout was in need of a better structure.

The "Add New" button (the dashed box with the cross in the middle) would always be added to the end of the items because it was inside the same container. This meant that users would have to scroll to the botom of the saved items to get to the button. Now, it has been taken out of the container and positioned so that it is always on the right, no matter how many items are saved.

The container with the items not has a max-height and scrolls, so that the page doesn't distort when more items are added. They are also now left-justifued, so that they display in a more ordered manner.

lastly, the banner containing the logo (and possbly the accounts icon in the future) shared the width of the items container, meaning it would move closer to the middle of the screen if there were 3 or less saved items. This looked bad, so it's been fixed to 100% width at all times.