fix(frontend): Update fields that are unique
Before, when we copied a user, it failed as the email and username
need to be unique. I therefore update those values in the copyUser
function. There is still the limitation that you can only copy a
user once, as it will then fail the unique values, as the email and
usernames will match existing user's.