fix(frontend): Fix annotations icons
I had to change the structure of the DropdownButton for the line width and line height, because usign an image with the icon as the src worked fine for an SVG, but not a React component. The styling was all broken, so I've fixed that, but it's something we will have to settle for good at some point, as there are a lot of conflicting paddings/heights, and the library's default styles take precedence over the app's styles.