fix(frontend): Fix selecting AOI
The form submit button to search for satellite scenes worked if you
loaded an existing search, but not if you select a new AOI. This is
now working again. One thing to note though is that it disappears if
you select another toolbar button, but is still there, when you go
back.

IssueID #461