fix(frontend): Added and styled storage meter
Added in the storage meter and it's icons/text, and styled to match designs. As with the buttons, we are awaiting the icons being designed, but they are sized and positioned, so they will just have to be switched out when the time comes.

Also, the strage meter has a fixed width and does not take classes, which will need to be resolved, as it is forcing the side-menu to widen to it's fixed length.