fix(backend): Update backend dependencies again
I didn't properly test this first time around, when I initially
upgraded to the latest `django-astrosat-users`. It turns out that
a previous update to that library, that hadn't yet been integrated
here, needs a new dependency `Pillow`. I have since updated that
libraries dependencies and updated this app to use that newer version.