fix(frontend): Update customer state in Redux, update proxy data
(First off, I updated the proxy clightly, adding in null fields to licences with no users for consistency).

I had to use a filter to remove the customer's customer_user IDs if the user with that ID is deleted. This proved to be more difficult than I thought it would be.

I tried several methods, the most promising being extracting into a helper function, but the importing exporting between React and Express are different, which made this impossible. I ended up duplicating the function, which isn't ideal, but is the only way I can see.

It's only a loop and an if, so not a huge duplication, and an acceptabole exchange for what it achieves.