fix(frontend): Revert slice to PUT
I originally added a PATCH method to the admin slice, but aparently it is a PUT request that will be used to update users.
I have also built the `editUser` action and proxy endpoint into this commit, because although it is changing a role, it is using the same endpoint as editing a user, and it seemed sinsible to simply re-use this action in both cases.