feat(backend): special handling of "core" Orb
Added a data migration ensuring that the "core" Orb exists.  Added
signals to automatically add/remove a Licence to the "core" Orb as
CustomerUsers are added/removed from Customers.  Added a couple of
LicenceQuerySet methods to filter the "core" Orb.  And ensured that
references to licences to the "core" Orb are excluded from the output of
serializers for Customers & CustomerUsers, while being explicitly added
to the input of those serializers.

IssueID #582