feat(backend): allowed core orb to be public
added is_private and is_core fields on Orb;  An orb is no longer
determined to be the "core" orb based on its name.  Only a single orb
can be the core orb so a constraint was added and a custom  clean method
on the admin form.  The LicenceQuerySet methodds were updated to return
public/private licences based on those fields instead of the name of the
orb.

IssueID #684