fix(frontend): Fix adding the one and only licence
When adding the one and only licence, we expected the hook to return
an array, like it was for when there are many licences. This was a
wrong assumption. I now check and if it is an array, use it, otherwise
wrap the value in an array and overwrite.

IssueID #707