fix(frontend): Improve cluster event handling
The feature popup was appearing when hovering over a cluster. The
level of data we have here makes it useless, so I prevent the popup
if hovering over a cluster.

When clicking a cluster, it should zoom in to show the first actual
feature, so mimicked what was done for `hourglass`.

IssueID #711