fix(frontend): Fix cluster icon crop
I made the width of the `iconAtlas` the exact width of the 2 icons inside it (29 + 88), then offset the yellow circle's x-axis to be the width of the pin icon plus half its own width (29 + 44).

This is due to the yellow circle icon rendering at its own center line, no the edge.