fix(frontend): show correct admin when users are members of multiple customers
If the user was a manager of any customer, the Admin Panel would naively
open the first customer of that user.  But that could be a different
customer or even a customer that the user is only a member - and not a
manager - of.

This PR simply returns the first customer that the user is a manager of.
It therefore assumes that a user is only a manager of one customer at
once.  Figuring out what to do when this is _not_ the case will be
handled in a future PR.