fix(backend): UserAdmin refers to an invalid CustomerAdmin
In order to add licences in bulk to a `Customer` in the admin, I created
a new admin for the `LicencedCustomer` _proxy_ model.  However, the
`UserAdmin` still had a reference to the old (now unused) `CustomerAdmin`.

Because `UserAdmin` was part of the django-astrosat-users, I couldn't
just change it.  Instead, I remapped the old Customer model to the new
LicencedCustomer Admin.

This has the added advantage of keeping the Customer Admin in the "astrosat-users" section, which is where it should be.