fix(frontend): Add constants
I switched the string identifiers for the user type to be constants imported form `admin.constants.js`, and made local constants for the button labels.

NOTE: Someone looking at this commit might wonder why I didn't just use the same constants inboth cases, the reasong being that the constants used for the object type ('MANAGER'/'MEMBER') is what the backend is expecting as data, whereas the labels ('Admin'/'Standard') is what the desigs say the buttons should display.

So the two choices were either to use use strings, or local constants, as the constants in `admin.constants.js` can not be changed.