fix(frontend): Limit potential for clashes on window object
I have introduced a new global object on the `window` object. This
does have the possibility to clash though, with 3rd party libs that
do the same. I have limited the potential for this by namespacing
our usage within `window.orbis`, this should make it very unlikely
that any clashes happen, but doesn't erradicate it, that we cannot
do using this solution, but neither do I want to over-engineer this
either.

IssueID #702