feat(frontend): Clean up login and register flows
The flow for a user registering and getting verified and the different
views showing was built into the login and register forms.
It was managed by little bits of state everywhere and sometimes required
2 or 3 re-renders to complete.
This moves the routing and logic into Redux and out of the components. Much neater
and facilitates component reuse better

IssueID #814