fix(frontend): Move max-width from licences to form
The max-width was originally on the `form` element, but didn't allow the it to fill the width of the form.

As the dialog was built to fit the size of the content passed in, I set a max-width on the form element instead (and shared it with the edit form for consistency).