feat(backend): add more email templates
Add email templates for assigning/revoking admin priveledges, updating
users, and updating customers.

Note that I had to refactor the to_representation method of the
CustomerSerializer b/c it was using a generator (filterfalse) and since
it is now being evaluated twice the generator would be empty the second time
around.

IssueID #829