fix(frontend): Add minimal user response object to global state, update tests.
I passed the returned streamlined user object to the `passwordResetConfirmSuccess` action, and added it to global state so that it can be used to determine wihether a user resetting their password is a new `teamMember` onboarding, or an existing user genuinely resettig their password.

Also updated both `login` and `admin.slice` tests to reflect these changes.