feat(backend): get property description from application-specific metadata
Get the description to use for the properties in the sidebar from
`application.orbis.description` if it exists.  If not, continue using
`description`.  This is particularly relevant for grouped properties; it
allows us to overwrite the default description w/ one that is applicable
to all members of the group.

IssueID #923